+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.0.0-rc.35",
-  genVersion: "2.753.7",
-  userAgent: "speakeasy-sdk/typescript 4.0.0-rc.35 2.753.7 2.0 @orq-ai/node",
+  sdkVersion: "4.0.0-rc.36",
+  genVersion: "2.755.0",
+  userAgent: "speakeasy-sdk/typescript 4.0.0-rc.36 2.755.0 2.0 @orq-ai/node",
 } as const;