+++ packages/orq-rc/src/models/operations/retrievedatasource.ts
@@ -145,7 +145,7 @@ export const RetrieveDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KA1CXJNFXHHXNKMB95R13RNG"),
+  _id: z.string().default("01KA3CD2J6JXJ8YMW6S31K2SNH"),
   display_name: z.string(),
   description: z.string().optional(),
   status: RetrieveDatasourceStatus$inboundSchema,
@@ -188,7 +188,7 @@ export const RetrieveDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   RetrieveDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01KA1CXJNFXHHXNKMB95R13RNG"),
+  id: z.string().default("01KA3CD2J6JXJ8YMW6S31K2SNH"),
   displayName: z.string(),
   description: z.string().optional(),
   status: RetrieveDatasourceStatus$outboundSchema,